update argcheck for eig loss, and remove repeat code in default dataset, rename tp.py
