+++ dptb/nn/tensor_product.py
